% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\docType{data}
\name{data_matrix}
\alias{data_matrix}
\title{0-1 value matrix of common cell types with their classical marker features}
\format{
A data frame with 157 columns and 17 rows
}
\usage{
data_matrix
}
\description{
A 0-1 value data.frame used for common cellular similarity corrleation calculation
}
\keyword{datasets}
