% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_querydata_preprocess.R
\name{get_cor_pre_sc_mmu_df}
\alias{get_cor_pre_sc_mmu_df}
\title{get_cor_pre_sc_mmu_df}
\usage{
get_cor_pre_sc_mmu_df(seurat)
}
\arguments{
\item{seurat}{query data seurat object}
}
\value{

}
\description{
Get_cor_pre_sc_mmu_df for mouse query data of healthy or cancer tissues at single cell-level
}
