% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predigs_prediction.R
\name{get_predigs_label_table}
\alias{get_predigs_label_table}
\title{get_predigs_label_table()}
\usage{
get_predigs_label_table(query_predigs_score_list, clusterlevel = TRUE)
}
\arguments{
\item{query_predigs_score_list}{a list of query-reference cell similarity score matrices after predigs prediction}

\item{clusterlevel}{TRUE or FALSE, which TRUE means cell type annotation at cluster-level ,FALSE means annotation at single cell-level}
}
\value{

}
\description{
get predicted label table after predigs cell prediction
}
\details{
predicted cellular label output at cluster or single cell level
}
\examples{
query_predigs_label_table_list <- get_predigs_label_table(query_predigs_score_list,clusterlevel = FALSE)
}
