% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predigs_prediction.R
\name{get_predigs_label_table_cluster}
\alias{get_predigs_label_table_cluster}
\title{get_predigs_label_table_cluster}
\usage{
get_predigs_label_table_cluster(query_predigs_score_list)
}
\arguments{
\item{query_predigs_score_list}{a list of query-reference cell similarity score matrices after predigs prediction}
}
\value{

}
\description{
get label table of predigs grouped cell prediction

predicted cellular label output at cluster-level
}
