% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predigs_prediction.R
\name{get_predigs_label_table_sc}
\alias{get_predigs_label_table_sc}
\title{get_predigs_label_table_sc}
\usage{
get_predigs_label_table_sc(query_predigs_score_list)
}
\arguments{
\item{query_predigs_score_list}{a list of query-reference cell similarity score matrices after predigs prediction}
}
\value{

}
\description{
get label table of predigs single cell prediction

predicted cellular label output at single cell-level
}
