% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predigs_prediction.R
\name{get_predigs_label_visualization}
\alias{get_predigs_label_visualization}
\title{get_predigs_label_visualization}
\usage{
get_predigs_label_visualization(query_predigs_score_df)
}
\arguments{
\item{query_predigs_score_df}{the data frame of query-reference cell similarity score matrix with predigs prediction}
}
\description{
The visualization of query-reference cell similarity score matrix with predigs prediction
}
\examples{
get_predigs_label_visualization(query_predigs_score_list[[1]])
}
