% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predigs_prediction.R
\name{get_predigs_label_visualization_sankey}
\alias{get_predigs_label_visualization_sankey}
\title{get_predigs_label_visualization_sankey}
\usage{
get_predigs_label_visualization_sankey(
  df,
  GroupNumberBalance = FALSE,
  cols = NULL
)
}
\arguments{
\item{df}{the data frame of querygroup-reference cell label data frame after predigs prediction}

\item{GroupNumberBalance}{choose whether viualize the querygroup-reference releationship in a balanced query-group number ,with FALSE at default.}

\item{cols}{a vector of color palettes ,with NULL at default}
}
\description{
The visualization of querygroup-reference cell label after predigs prediction
}
