% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predigs_prediction.R
\name{get_predigs_score_df}
\alias{get_predigs_score_df}
\title{get_predigs_score_df()}
\usage{
get_predigs_score_df(reference_cor_list, querydata_cor_list, i)
}
\arguments{
\item{reference_cor_list}{a list of multi-reference datasets for cell type assignment from predigs}

\item{querydata_cor_list}{a list of multi-query datasets, and a query list with only one dataset is allowed}

\item{i}{the index value of querydata_cor_list, i.e., it could be 1,2,3,4,5 if the length of querydata_cor_list is equal to 5}
}
\value{

}
\description{
cell type prediction through predigs
}
\details{
calculate gene expression similarity matrix between query cells and reference cell types based on the multiple datasets integration
}
\examples{
data("ref_health_pancrea_cor_list")
querydata_list <- list(query_df)
query_predigs_score_list <- lapply(1:length(querydata_list),function(x){
    get_predigs_score_df(reference_cor_list = ref_health_pancrea_cor_list,
                                         querydata_cor_list = querydata_list,
                                         i=x)
})
}
