% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_querydata_preprocess.R
\name{get_query_preprocess}
\alias{get_query_preprocess}
\title{get_query_preprocess()}
\usage{
get_query_preprocess(seurat, species = "human", cellgroup, clusterlevel = TRUE)
}
\arguments{
\item{seurat}{a query seurat object}

\item{species}{the species of query scRNAseq data ,i.e.,human or mouse}

\item{cellgroup}{the column name of cell type labels in the metadata data frame of each query seurat object}

\item{clusterlevel}{TRUE or FALSE, which TRUE means cell type annotation at cluster-level ,FALSE means annotation at single cell-level}
}
\value{

}
\description{
Query Data Preprocessing Workflow for PreDigs
}
\examples{
query_df <- get_query_preprocess(alldata,species = "human",cellgroup="cell_ontology_class",clusterlevel = TRUE)

}
