% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference_setname.R
\name{get_reference_setname}
\alias{get_reference_setname}
\title{predigs reference datasets list loading}
\usage{
get_reference_setname(organtype = NULL, tissuetype = NULL)
}
\arguments{
\item{organtype}{specific organ type ,i.e.,intestine/pancrea/stomach/liver/esophagus}

\item{tissuetype}{specific tissue type ,i.e.,cancer/health}
}
\value{

}
\description{
load a list of multiple reference datasets
}
\details{
The reference list for sepcific organ with specific tissue types in the human digestive system
}
\examples{
reference_dataset <- get_reference_setname("liver","health")
}
