% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seurat_workflow.R
\name{get_seurat_workflow}
\alias{get_seurat_workflow}
\title{seurat workflow started with raw seurat count object}
\usage{
get_seurat_workflow(RawCountSeurat)
}
\arguments{
\item{RawCountSeurat}{seurat object without "data" slot in the RNA assays}
}
\value{

}
\description{
preprocess of usual seurat workflow
}
\details{
\preformatted{ The Workflow includes:
 (1) Normalization[Seurat::NormalizeData;FindVariableFeatures;ScaleData;RunPCA]
 (2) Batch Effect Removal[harmony::RunHarmony]
 (3) Clustering--[Seurat::RunUMAP;FindNeighbors;FindClusters;RunTSNE]
}
}
