% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\docType{data}
\name{my36colors}
\alias{my36colors}
\title{A vector of 36 common color palette value used for ggplot2 visualization}
\format{
An object of class \code{character} of length 36.
}
\source{
\url{https://www.jianshu.com/p/db0d0927e613}
}
\usage{
my36colors
}
\description{
Hexadecimal value
}
\keyword{datasets}
