% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_cancer_stomach_cor_list}
\alias{ref_cancer_stomach_cor_list}
\title{reference_cor_list of sepcific digestive organ types in tumor tissue}
\format{
An object of class \code{list} of length 5.
}
\usage{
ref_cancer_stomach_cor_list
}
\description{
A list of multiple data frame used for cellular similarity corrleation calculation
}
\keyword{datasets}
