% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\docType{data}
\name{reference_gene}
\alias{reference_gene}
\title{reference_genes of digestive cells}
\format{
A data frame with 1 columns:
\describe{
\item{Gene}{reference genes symbol or gene name}
}
}
\usage{
reference_gene
}
\description{
A data frame of genes used for cellular similarity corrleation calculation
}
\keyword{datasets}
