% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\docType{data}
\name{trans}
\alias{trans}
\title{human-mouse gene conversion data frame}
\format{
A data frame with 75483 columns and 4 rows
}
\source{
\url{http://asia.ensembl.org/biomart/martview/65b916eaf1c6cf8f56e0ad08abf2a114}
}
\usage{
trans
}
\description{
A data frame with matched gene between human-hsa and mouse-mmu
}
\keyword{datasets}
